+++ packages/analytics/analytics-chart/src/enums/chart-types-simple.enum.ts
@@ -1,5 +1,5 @@
-/* eslint-disable no-unused-vars */
+
 export enum ChartTypesSimple {
   GAUGE = 'Gauge',
-  TOPN = 'TopN'
+  TOPN = 'TopN',
 }