+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.22.1...@kong-ui-public/dashboard-renderer@0.22.2) (2024-06-07)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.22.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.22.0...@kong-ui-public/dashboard-renderer@0.22.1) (2024-06-07)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+# [0.22.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.21.28...@kong-ui-public/dashboard-renderer@0.22.0) (2024-06-06)
+
+
+### Features
+
+* **eslint:** eslint v9 and shared configs [KHCP-11627] ([#1435](https://github.com/Kong/public-ui-components/issues/1435)) ([d0a973f](https://github.com/Kong/public-ui-components/commit/d0a973ff61b5302ee2b1cd3d73b0deb0c5864fa5))
+
+
+
+
+
 ## [0.21.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.21.27...@kong-ui-public/dashboard-renderer@0.21.28) (2024-06-03)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer